requirements.txt
Flask
Flask-SQLAlchemy
pandas
requests
python-dotenv